fix: only show context menu on details page
I just used the standard GetEntities query which uses
minimal[entity-type] for its definition. I did this because i want to
avoid a lot of copy-pasting and duplicated work. The only drawback is
that, of course, this way the context menu is also displayed on the
overview page etc. This commits prevents that.